+++ Backend/BOSS/api/serializer.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from .models import Message
+
+class MessageSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Message
+        fields = ['text']
